+++ service/src/main/kotlin/io/provenance/explorer/grpc/v1/MarkerGrpcClient.kt
@@ -57,5 +57,16 @@ class MarkerGrpcClient(channelUri: URI) {
             QueryHoldingResponse.getDefaultInstance()
         }
 
+    fun getAllMarkerHolders(denom: String): QueryHoldingResponse =
+        try {
+            markerClient.holding(
+                QueryHoldingRequest.newBuilder()
+                    .setId(denom)
+                    .build()
+            )
+        } catch (e: Exception) {
+            QueryHoldingResponse.getDefaultInstance()
+        }
+
     fun getMarkerParams() = markerClient.params(MarkerRequest.newBuilder().build())
 }